Added collision resolution

+++ FlashlightTag2D/FlashlightTag2D/AIComponent.cpp
@@ -15,7 +15,7 @@ AIComponent::~AIComponent()
 {
 }
 
-void AIComponent::Update(GameActor& actor, int deltaMs)
+void AIComponent::Update(GameActor& actor, float deltaMs)
 {
 	// If hiding
 	//  Look for seeker in light cone
