+++ FlashlightTag2D/FlashlightTag2D/AIComponent.h
@@ -10,7 +10,7 @@ public:
 	AIComponent();
 	virtual ~AIComponent();
 
-	void Update(GameActor& actor, int deltaMs) override;
+	void Update(GameActor& actor, float deltaMs) override;
 
 	ComponentId GetComponentId() const;
 	virtual EComponentNames GetComponentName() const override;
