+++ FlashlightTag2D/FlashlightTag2D/FollowTargetAIComponent.h
@@ -10,7 +10,7 @@ public:
 	FollowTargetAIComponent(StrongGameActorPtr target);
 	virtual ~FollowTargetAIComponent();
 
-	void Update(GameActor& actor, int deltaMs) override;
+	void Update(GameActor& actor, float deltaMs) override;
 
 	ComponentId GetComponentId() const;
 	virtual EComponentNames GetComponentName() const override;
