+++ FlashlightTag2D/FlashlightTag2D/GraphicsComponent.cpp
@@ -28,40 +28,17 @@ EComponentNames GraphicsComponent::GetComponentName() const
 	return EComponentNames::GraphicsComponentEnum;
 }
 
-//std::vector<float> GraphicsComponent::GetScaledVertices(int screenWidth, int screenHeight)
-//{
-//	std::vector<float> vertices;
-//
-//	if (screenHeight < 1 || screenWidth < 1)
-//	{
-//		return vertices;
-//	}
-//
-//	auto pos = m_pTransformComponent->GetPosition();
-//
-//	// Top Right
-//	vertices.push_back((pos.x + m_imageOffset.x)/screenWidth); vertices.push_back((pos.y - m_imageOffset.y) / screenHeight); vertices.push_back(0);
-//	// Bottom Right
-//	vertices.push_back((pos.x + m_imageOffset.x) / screenWidth); vertices.push_back((pos.y + m_imageOffset.y) / screenHeight); vertices.push_back(0);
-//	// Bottom Left
-//	vertices.push_back((pos.x - m_imageOffset.x) / screenWidth); vertices.push_back((pos.y + m_imageOffset.y) / screenHeight); vertices.push_back(0);
-//	// Top Left 
-//	vertices.push_back((pos.x - m_imageOffset.x) / screenWidth); vertices.push_back((pos.y - m_imageOffset.y) / screenHeight); vertices.push_back(0);
-//
-//	return vertices;
-//}
-
-void GraphicsComponent::Update(GameActor& actor, int deltaMs)
+void GraphicsComponent::Update(GameActor& actor, float deltaMs)
 {
 	if (m_curAnimationTime < 0)
 	{
 		m_curAnimationTime = m_animationTimer;
 
-		m_texturePos.x = (int)(m_texturePos.x + m_pTransformComponent->GetSize().x) % m_texture.GetWidth();
+		m_texturePos.x = (float)((int)(m_texturePos.x + m_pTransformComponent->GetSize().x) % m_texture.GetWidth());
 		
 		if (m_texturePos.x == 0)
 		{
-			m_texturePos.y = (int)(m_texturePos.y - m_pTransformComponent->GetSize().y) % m_texture.GetHeight();
+			m_texturePos.y = (float)((int)(m_texturePos.y - m_pTransformComponent->GetSize().y) % m_texture.GetHeight());
 		}
 	}
 	else
