+++ FlashlightTag2D/FlashlightTag2D/InputComponent.h
@@ -8,7 +8,7 @@ public:
 	InputComponent();
 	virtual ~InputComponent();
 
-	void Update(GameActor& actor, int deltaMs) override;
+	void Update(GameActor& actor, float deltaMs) override;
 
 	ComponentId GetComponentId() const;
 	virtual EComponentNames GetComponentName() const override;
